feat(admin,api): 2381 - [PDT] identification des lignes fusionnees (#3848)
* feat(admin,api): 2381 - [PDT] identification des lignes fusionnees

* fix con

* sync mergedLines on new import, fix export column names, add existing bus id control

* fix ts checks

* feedbacks

* types / naming

* fix import

* fix import
